% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.R
\name{ace_module_ui}
\alias{ace_module_ui}
\title{ACE Editor Module UI}
\usage{
ace_module_ui(id, exprs_init = NULL)
}
\arguments{
\item{id}{A unique identifier for the module's UI.}

\item{exprs_init}{Optional initial expression(s) to load into the editor.}
}
\value{
A Shiny UI element for the ACE Editor module.
}
\description{
This function creates a user interface for the ACE Editor module in a Shiny application.
It can be used to dynamically add and manage ACE Editor instances.
}
\examples{
\dontrun{
ui <- bslib::page_fluid(
  ace_module_ui("m1", exprs_init = c(a = "bla", b = "blabla")),
  verbatimTextOutput("o_result")
)
server <- function(input, output) {
  r_result <- ace_module_server("m1")
  output$o_result <- renderPrint({
    r_result()
  })
}
# Run the application
shinyApp(ui = ui, server = server)
}
}
