% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{add_block}
\alias{add_block}
\title{Add block to a stack}
\usage{
add_block(stack, block, position = NULL)
}
\arguments{
\item{stack}{stack to update. See \link{new_stack}.}

\item{block}{Block to insert.}

\item{position}{Where to insert the new block. If NULL,
the block will be added in the last position.}
}
\value{
Invisibly returns the stack.
}
\description{
This is to be called oustide the stack by
other modules.
}
