% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{add_block_server}
\alias{add_block_server}
\title{Add block server module
This modules aims at showing extra info in the
offcanvas menu to add blocks. Blocks are added
at the stack level with another function.}
\usage{
add_block_server(id, vals)
}
\arguments{
\item{id}{Module id.}

\item{vals}{Reactive values.}
}
\description{
Add block server module
This modules aims at showing extra info in the
offcanvas menu to add blocks. Blocks are added
at the stack level with another function.
}
