% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialise.R
\name{as_list}
\alias{as_list}
\alias{as_list.field}
\alias{as_list.block}
\alias{as_list.stack}
\title{As list
Convert stacks, blocks, and fields to \code{list}.}
\usage{
as_list(x, ...)

\method{as_list}{field}(x, ...)

\method{as_list}{block}(x, ...)

\method{as_list}{stack}(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Ignored.}
}
\description{
As list
Convert stacks, blocks, and fields to \code{list}.
}
\section{Methods (by class)}{
\itemize{
\item \code{as_list(field)}: Convert a field

\item \code{as_list(block)}: Convert a block

\item \code{as_list(stack)}: Convert a block

}}
