% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{block_body}
\alias{block_body}
\alias{block_body.block}
\title{Block body generic}
\usage{
block_body(x, ...)

\method{block_body}{block}(x, ns, inputs_hidden, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{ns}{Module namespace}

\item{inputs_hidden}{For styling purposes: CSS class to apply
when the block is collapsed.}
}
\description{
Generic for creating fields UI container
}
