% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{block_code}
\alias{block_code}
\alias{block_code.block}
\title{Block code generic}
\usage{
block_code(x, ...)

\method{block_code}{block}(x, ns, inputs_hidden, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{ns}{Module namespace}

\item{inputs_hidden}{For styling purposes: CSS class to apply
when the block is collapsed.}
}
\description{
Generic for creating the block code UI elements that
are the toggle and the code container.
}
