% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{block_combiner}
\alias{block_combiner}
\alias{block_combiner.transform_block}
\alias{block_combiner.data_block}
\alias{block_combiner.plot_block}
\alias{block_combiner.plot_layer_block}
\title{Combine 2 block expressions}
\usage{
block_combiner(x, ...)

\method{block_combiner}{transform_block}(x, y, ...)

\method{block_combiner}{data_block}(x, y, ...)

\method{block_combiner}{plot_block}(x, y, ...)

\method{block_combiner}{plot_layer_block}(x, y, ...)
}
\arguments{
\item{x}{Block object.}

\item{...}{For generic consistency.}

\item{y}{Block object.}
}
\description{
Useful for \link{generate_code}.
}
