% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{block_download_ui}
\alias{block_download_ui}
\alias{block_download_ui.default}
\alias{block_download_ui.transform_block}
\alias{block_download_ui.data_block}
\alias{block_download_ui.plot_block}
\title{Block download generic}
\usage{
block_download_ui(x, ...)

\method{block_download_ui}{default}(x, ns, inputs_hidden, ...)

\method{block_download_ui}{transform_block}(x, ns, inputs_hidden, ...)

\method{block_download_ui}{data_block}(x, ns, inputs_hidden, ...)

\method{block_download_ui}{plot_block}(x, ns, inputs_hidden, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{ns}{Module namespace}

\item{inputs_hidden}{For styling purposes: CSS class to apply
when the block is collapsed.}
}
\description{
Generic to create the block download button.
}
