% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{block_header}
\alias{block_header}
\alias{block_header.block}
\title{Block header generic}
\usage{
block_header(x, ...)

\method{block_header}{block}(x, ns, hidden_class, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{ns}{Module namespace}

\item{hidden_class}{TBD.}
}
\description{
Generic for creating the block header.
}
