% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{block_icon}
\alias{block_icon}
\alias{block_icon.default}
\alias{block_icon.data_block}
\alias{block_icon.transform_block}
\alias{block_icon.plot_block}
\title{Block icon generic}
\usage{
block_icon(x, ...)

\method{block_icon}{default}(x, ...)

\method{block_icon}{data_block}(x, ...)

\method{block_icon}{transform_block}(x, ...)

\method{block_icon}{plot_block}(x, ...)
}
\arguments{
\item{x}{Object inheriting from \code{"block"}.}

\item{...}{For generic consistency.}
}
\description{
Create a block icon depending in the block class
}
