% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-core.R
\name{block_input_check}
\alias{block_input_check}
\alias{block_input_check.data_block}
\alias{block_input_check.transform_block}
\alias{block_input_check.parser_block}
\alias{block_input_check.result_block}
\alias{input_failure}
\alias{block_output_ptype}
\alias{block_output_ptype.dataset_block}
\alias{block_output_ptype.result_block}
\alias{block_output_ptype.upload_block}
\alias{block_output_ptype.filesbrowser_block}
\alias{block_output_ptype.transform_block}
\title{Block input/output}
\usage{
block_input_check(x, data, ...)

\method{block_input_check}{data_block}(x, data, ...)

\method{block_input_check}{transform_block}(x, data, ...)

\method{block_input_check}{parser_block}(x, data, ...)

\method{block_input_check}{result_block}(x, data, ...)

input_failure(..., class = character())

block_output_ptype(x, ...)

\method{block_output_ptype}{dataset_block}(x, ...)

\method{block_output_ptype}{result_block}(x, ...)

\method{block_output_ptype}{upload_block}(x, ...)

\method{block_output_ptype}{filesbrowser_block}(x, ...)

\method{block_output_ptype}{transform_block}(x, ...)
}
\arguments{
\item{x}{Block}
}
\description{
Used for checking whether blocks are compatible
}
