% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr-package}
\alias{blockr}
\alias{blockr-package}
\title{blockr: A Block-Based Framework for Data Manipulation and Visualization}
\description{
A framework for building interactive dashboards and document-based reports. Underlying data manipulation and visualization is possible using a web-based point and click user interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bristolmyerssquibb.github.io/blockr/}
}

}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

Authors:
\itemize{
  \item David Granjon \email{david@cynkra.com}
  \item Christoph Sax \email{christoph@cynkra.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})
  \item John Coene \email{john@opifex.org}
}

Other contributors:
\itemize{
  \item Karma Tarap \email{Karma.Tarap@bms.com} [contributor]
  \item  Bristol Myers Squibb [copyright holder, funder]
}

}
\keyword{internal}
