% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{blockr_serialize}
\alias{blockr_serialize}
\alias{blockr_serialize.field}
\alias{blockr_serialize.block}
\alias{blockr_serialize.stack}
\alias{blockr_serialize.workspace}
\alias{blockr_deserialize}
\alias{from_json}
\alias{to_json}
\title{Serialization}
\usage{
blockr_serialize(x)

\method{blockr_serialize}{field}(x)

\method{blockr_serialize}{block}(x)

\method{blockr_serialize}{stack}(x)

\method{blockr_serialize}{workspace}(x)

blockr_deserialize(x)

from_json(x)

to_json(x = get_workspace())
}
\arguments{
\item{x}{Object to (de)serialize}
}
\description{
Object (de)serialization.
}
