% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{can_compute}
\alias{can_compute}
\alias{can_compute.block}
\alias{can_compute.select_block}
\alias{can_compute.arrange_block}
\alias{can_compute.group_by_block}
\alias{can_compute.filter_block}
\alias{can_compute.join_block}
\alias{can_compute.summarize_block}
\title{Allow/Prevent computation}
\usage{
can_compute(x, ...)

\method{can_compute}{block}(x, iv, input, in_dat, ...)

\method{can_compute}{select_block}(x, iv, input, in_dat, ...)

\method{can_compute}{arrange_block}(x, iv, input, in_dat, ...)

\method{can_compute}{group_by_block}(x, iv, input, in_dat, ...)

\method{can_compute}{filter_block}(x, iv, input, in_dat, ...)

\method{can_compute}{join_block}(x, iv, input, in_dat, ...)

\method{can_compute}{summarize_block}(x, iv, input, in_dat, ...)
}
\arguments{
\item{x}{Block.}

\item{...}{Generic consistency}

\item{iv}{Input validator (R6 instance).}

\item{input}{Input name to validate. Without namespace.}

\item{in_dat}{Reactive input data}
}
\value{
Boolean value depending on the state.
}
\description{
Each block is able to validate the input state
to see whether values are compatible with the incoming
data to process.
}
