% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{cleanup_block}
\alias{cleanup_block}
\title{Cleanup module inputs}
\usage{
cleanup_block(x, input, output, obs)
}
\arguments{
\item{x}{Block.}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{obs}{Module observer list.}
}
\description{
Cleanup module inputs
}
\keyword{internal}
