% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_block}
\alias{convert_block}
\title{Convert block from a type to another}
\usage{
convert_block(from = new_select_block, to, data, ...)
}
\arguments{
\item{from}{Block function to start from like new_select_block.}

\item{to}{dplyr verb (function, not a string!) such as arrange, group_by...}

\item{data}{Necessary to \link{initialize_block}.}

\item{...}{Necessary to \link{initialize_block}.}
}
\description{
For instance, you can convert from a select block to an
arrange block or group_by which have similar structure.
}
\keyword{internal}
