% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_block}
\alias{create_block}
\title{Create a new block}
\usage{
create_block(name, type = c("transform", "plot"), file = NULL)
}
\arguments{
\item{name}{Block name.}

\item{type}{Type of block to create.}

\item{file}{Output file.
If \code{NULL} then it constructs \verb{R/<name>-block.R}.}
}
\description{
Create a new block
}
