% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_data_block}
\alias{new_data_block}
\alias{data_block}
\title{Data block constructor}
\usage{
new_data_block(
  ...,
  dat = as.environment("package:datasets"),
  selected = character()
)

data_block(...)
}
\arguments{
\item{...}{Further (metadata) attributes}

\item{dat}{Multiple datasets.}

\item{selected}{Selected dataset.}
}
\description{
This block allows to selected data from a package,
the default being datasets.
}
