% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{data_info}
\alias{data_info}
\alias{data_info.block}
\alias{data_info.data_block}
\alias{data_info.transform_block}
\title{Data info generic}
\usage{
data_info(x, ...)

\method{data_info}{block}(x, ns, ...)

\method{data_info}{data_block}(x, ns, ...)

\method{data_info}{transform_block}(x, ns, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}
}
\description{
Generic for creating the data info tags,
to display the number of columns, rows, ...
}
