% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{default_layout_fields}
\alias{default_layout_fields}
\title{Default layout for fields
Default layout for fields, places fields on rows
and collapses parts of inputs when there are too many.}
\usage{
default_layout_fields(fields, ...)
}
\arguments{
\item{fields}{List of fields.}

\item{...}{Currently ignored.}
}
\description{
Default layout for fields
Default layout for fields, places fields on rows
and collapses parts of inputs when there are too many.
}
\keyword{internal}
