% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_ui}
\alias{download_ui}
\alias{download_ui.block}
\alias{download_ui.transform_block}
\alias{download_ui.stack}
\alias{download_ui.data_block}
\alias{download_ui.plot_block}
\title{Block download generic}
\usage{
download_ui(x, ...)

\method{download_ui}{block}(x, ns, inputs_hidden = FALSE, ...)

\method{download_ui}{transform_block}(x, ns, inputs_hidden = FALSE, ...)

\method{download_ui}{stack}(x, ns, inputs_hidden = FALSE, ...)

\method{download_ui}{data_block}(x, ns, inputs_hidden = FALSE, ...)

\method{download_ui}{plot_block}(x, ns, inputs_hidden = FALSE, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{ns}{Module namespace}

\item{inputs_hidden}{For styling purposes: CSS class to apply
when the block is collapsed.}
}
\description{
Generic to create the block download button.
}
