% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-core.R, R/fields.R
\name{field_value}
\alias{field_value}
\alias{field_value.field}
\alias{field_component}
\alias{field_component.field}
\alias{field_value.variable_field}
\alias{field_value.list_field}
\alias{field_value.kv_field}
\title{Get field attribute value}
\usage{
field_value(x)

\method{field_value}{field}(x)

field_component(x, name)

\method{field_component}{field}(x, name)

\method{field_value}{variable_field}(x)

\method{field_value}{list_field}(x)

\method{field_value}{kv_field}(x)
}
\arguments{
\item{x}{An object inheriting form \code{"field"}}

\item{name}{Component name}
}
\value{
Field value
}
\description{
Get the field value attribute. If it is a function,
return the result attribute instead.
}
