% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_filter_block}
\alias{new_filter_block}
\alias{filter_block}
\title{Filter block}
\usage{
new_filter_block(
  data,
  columns = character(),
  values = character(),
  filter_fun = "==",
  ...
)

filter_block(data, ...)
}
\arguments{
\item{data}{Tabular data to filter (rows).}

\item{columns}{Definition of the equality filter.}

\item{values}{Definition of the equality filter.}

\item{filter_fun}{Default filter fun for the expression.}

\item{...}{Further (metadata) attributes}
}
\description{
This block provides access to \link[dplyr]{filter} verb and
returns the filtered data.
}
