% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_dep}
\alias{find_dep}
\title{Finds dependency}
\usage{
find_dep(tag, dep)
}
\arguments{
\item{tag}{Tag inside which to find the dependency.}

\item{dep}{Dependency regex pattern like
\verb{^bootstrap-5*}.}
}
\value{
Boolean. TRUE if dependency found.
}
\description{
Find dependency through pattern matching.
}
\keyword{internal}
