% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-core.R
\name{generate_code}
\alias{generate_code}
\alias{generate_code.block}
\alias{generate_code.arrange_block}
\alias{generate_code.group_by_block}
\alias{generate_code.transform_block}
\alias{generate_code.plot_block}
\alias{generate_code.data_block}
\alias{generate_code.call}
\title{Generate code generic}
\usage{
generate_code(x)

\method{generate_code}{block}(x)

\method{generate_code}{arrange_block}(x)

\method{generate_code}{group_by_block}(x)

\method{generate_code}{transform_block}(x)

\method{generate_code}{plot_block}(x)

\method{generate_code}{data_block}(x)

\method{generate_code}{call}(x)
}
\arguments{
\item{x}{An object inheriting from \code{"block"}}
}
\description{
For a given block, generate the code contained
in the \code{expr} attribute. Needed by \link{evaluate_block}
and \link{block_combiner} to generate the entire stack code.
}
