% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{generate_server}
\alias{generate_server}
\alias{generate_server.block}
\alias{generate_server.stack}
\title{UI}
\usage{
generate_server(x, ...)

\method{generate_server}{block}(x, id, ...)

\method{generate_server}{stack}(x, id, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{id}{UI IDs}
}
\description{
Generic for UI generation
}
