% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvalue_field.R, R/ui.R
\name{ui_input.keyvalue_field}
\alias{ui_input.keyvalue_field}
\alias{generate_ui}
\alias{generate_ui.block}
\alias{generate_ui.stack}
\alias{generate_ui.workspace}
\alias{input_ids.list_field}
\title{UI}
\usage{
\method{ui_input}{keyvalue_field}(x, id, name)

generate_ui(x, ...)

\method{generate_ui}{block}(x, id, ..., .hidden = !getOption("BLOCKR_DEV", FALSE))

\method{generate_ui}{stack}(x, id = NULL, ...)

\method{generate_ui}{workspace}(x, id, ...)

\method{input_ids}{list_field}(x, name, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{id}{UI IDs}

\item{...}{Generic consistency}

\item{.hidden}{Whether to initialise the block hidden.}
}
\description{
Generic for UI generation
}
