% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_compatible_blocks}
\alias{get_compatible_blocks}
\title{Find stack compatible blocks}
\usage{
get_compatible_blocks(stack)
}
\arguments{
\item{stack}{Stack object.}
}
\value{
a dataframe.
}
\description{
Given a stack, we use the registry to find
what are the blocks compatible with the last stack block.
If the stack is empy, we return data blocks.
}
