% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_registry}
\alias{get_registry}
\title{List available blocks as a data.frame}
\usage{
get_registry()
}
\value{
A dataframe.
}
\description{
Provides an alternate way of displaying
the registry information.
This can be useful to create dynamic UI elements
like in \link{add_block_ui}.
}
