% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{handle_remove}
\alias{handle_remove}
\alias{handle_remove.block}
\title{Remove stack/block generic}
\usage{
handle_remove(x, ...)

\method{handle_remove}{block}(x, vals, session = getDefaultReactiveDomain(), ...)
}
\arguments{
\item{x}{Object for which to generate a \code{\link[=moduleServer]{moduleServer()}}}

\item{...}{Generic consistency}

\item{vals}{Internal reactive values.}

\item{session}{Shiny session object.}
}
\description{
Generic for stack/block removal

Necessary to be able to remove the block from the stack.
}
