% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{handle_remove_block.block}
\alias{handle_remove_block.block}
\title{Attach an observeEvent to the given block}
\usage{
\method{handle_remove_block}{block}(x, vals, session = getDefaultReactiveDomain())
}
\description{
Necessary to be able to remove the block from the stack.
}
