% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{handle_remove_block}
\alias{handle_remove_block}
\title{Handle block generic}
\usage{
handle_remove_block(x, ...)
}
\arguments{
\item{x}{Block element.}

\item{...}{Generic consistency.}
}
\description{
Generic for block removal
}
