% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{init_stack_block}
\alias{init_stack_block}
\alias{init_stack_block.call}
\alias{init_stack_block.name}
\title{Instantiate a block within a stack}
\usage{
init_stack_block(x, position, ...)

\method{init_stack_block}{call}(x, position, ...)

\method{init_stack_block}{name}(x, position, ...)
}
\arguments{
\item{x}{May be a call or constructor (function).}

\item{position}{Block index (position in the stack).}

\item{...}{For generic consistency.}
}
\value{
The evaluated block
}
\description{
Allows to pass either constructors or
block calls to a stack. The idea is to be able
to do either \code{new_stack(data_block, select_block)} or
\code{new_stack(data_block(selected = "iris"), select_block(columns = "Sepal.Length"))}.
}
\keyword{internal}
