% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{init_block}
\alias{init_block}
\title{Init a single block}
\usage{
init_block(i, vals, stack_id)
}
\arguments{
\item{i}{Block position}

\item{vals}{Reactive values containing the stack}

\item{stack_id}{Stack ID}
}
\description{
Useful for init.blocks but also
to be called after \link{add_block}.
}
\keyword{internal}
