% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_input_validator}
\alias{init_input_validator}
\title{Initialize input validator}
\usage{
init_input_validator(x, iv, update = FALSE)
}
\arguments{
\item{x}{Current block.}

\item{iv}{R6 validator instance.}

\item{update}{FALSE to init. TRUE to update
and add new rules.}
}
\value{
Result or error message.
}
\description{
To validate inputs in server blocks.
}
\keyword{internal}
