% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{init_lock}
\alias{init_lock}
\title{Initialise Lock
Initialises the lock reactive value.}
\usage{
init_lock(session = shiny::getDefaultReactiveDomain(), ignore_init = FALSE)
}
\arguments{
\item{session}{Shiny session.}

\item{ignore_init}{Whether to ignore the initialisation.
this is only used when we need to initialise the lock from \code{toggle_lock()}.}
}
\description{
Initialise Lock
Initialises the lock reactive value.
}
