% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-core.R, R/field-core.R, R/s3-utils.R
\name{is_initialized.block}
\alias{is_initialized.block}
\alias{is_initialized.field}
\alias{is_initialized}
\title{S3 generic for initialization}
\usage{
\method{is_initialized}{block}(x)

\method{is_initialized}{field}(x)

is_initialized(x)
}
\arguments{
\item{x}{Element to check.}
}
\value{
Boolean value.
}
\description{
Checks if a block or field (currently implemented methods)
is initialized.
}
