% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-core.R
\name{initialize_field}
\alias{initialize_field}
\alias{initialize_field.field}
\title{Initialize field generic}
\usage{
initialize_field(x, env = list())

\method{initialize_field}{field}(x, env = list())
}
\arguments{
\item{x}{An object inheriting form \code{"field"}}

\item{env}{Environment with data and other field values}
}
\value{
The field.
}
\description{
TBD
}
