% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvalue_field.R
\name{generate_server.keyvalue_field}
\alias{generate_server.keyvalue_field}
\alias{ui_input.keyvalue_field}
\alias{new_keyvalue_field}
\alias{keyvalue_field}
\alias{validate_field.keyvalue_field}
\title{Re-implement numeric_field, using server module}
\usage{
\method{generate_server}{keyvalue_field}(x)

\method{ui_input}{keyvalue_field}(x, id, name)

new_keyvalue_field(value = numeric(), ...)

keyvalue_field(...)

\method{validate_field}{keyvalue_field}(x)
}
\description{
Re-implement numeric_field, using server module
}
