% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout}
\alias{layout}
\title{Default layout for fields
Default layout for fields, places fields on rows
and collapses parts of inputs when there are too many.}
\usage{
layout(x, fields, ...)
}
\arguments{
\item{x}{Object (block).}

\item{fields}{Named list of fields.}

\item{...}{Currently ignored.}
}
\description{
Default layout for fields
Default layout for fields, places fields on rows
and collapses parts of inputs when there are too many.
}
