% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_list_field}
\alias{new_list_field}
\alias{list_field}
\title{List field constructor}
\usage{
new_list_field(value = list(), sub_fields = list(), ...)

list_field(...)
}
\arguments{
\item{value}{Field value}

\item{sub_fields}{Fields contained in \code{list_field}}

\item{...}{Further field components}
}
\description{
A field that can contain subfields. See \link{slider_block} for
a usecase.
}
