% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{lock}
\alias{lock}
\alias{unlock}
\alias{toggle_lock}
\alias{is_locked}
\alias{observe_lock}
\title{Lock}
\usage{
lock(session = shiny::getDefaultReactiveDomain())

unlock(session = shiny::getDefaultReactiveDomain())

toggle_lock(session = shiny::getDefaultReactiveDomain())

is_locked(session = shiny::getDefaultReactiveDomain())

observe_lock(fn, ..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{Shiny session.}

\item{fn}{Callback function to run when the
editing locked/unlocked. It must accept a single
argument, which is a boolean indicating whether
the editing is locked or not.}

\item{...}{passed to \code{\link[shiny:observeEvent]{shiny::observeEvent()}}.}
}
\description{
Lock edit of stacks and blocks.
}
