% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-block.R
\name{new_mutate_block}
\alias{new_mutate_block}
\alias{mutate_block}
\title{Mutate block constructor}
\usage{
new_mutate_block(data, value = NULL, ...)

mutate_block(data, ...)
}
\arguments{
\item{data}{Tabular data in which to select some columns.}

\item{value}{Default value.}

\item{...}{Further (metadata) attributes}
}
\description{
Leverages the \link{keyvalue_field}
}
