% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_arrange_block}
\alias{new_arrange_block}
\title{Arrange block}
\usage{
new_arrange_block(columns = character(), ...)
}
\arguments{
\item{columns}{Column(s) to select.}

\item{...}{Further (metadata) attributes}
}
\description{
This block provides access to \link[dplyr]{arrange} verb and
returns a dataframe.
}
