% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-core.R
\name{new_block}
\alias{new_block}
\alias{is_block}
\title{Blocks}
\usage{
new_block(
  fields,
  expr,
  name = rand_names(),
  submit = FALSE,
  ...,
  class = character()
)

is_block(x)
}
\arguments{
\item{fields}{A list of field, each entry inheriting from \code{"field"}}

\item{expr}{A quoted expression (compatible with partial substitution as
implemented in \code{\link[base:bquote]{base::bquote()}} and intended for evaluation in the context
of the fields)}

\item{name}{Block name}

\item{submit}{Whether this block requires to press a submit button to get
the results. Boolean, default to FALSE, which means no submit button.
If NA, then no computation is triggered and the user needs to click on the button
to see the block result. If TRUE, computation is automatically triggered and the button
is shown (which is useful when restoring a stack).}

\item{...}{Further (metadata) attributes}

\item{class}{Block subclass}

\item{x}{An object inheriting from \code{"block"}}
}
\description{
Blocks consist of a set of fields, an expression that defines what the block
produces (given the result of the previous block combined with user input),
plus some meta data.
}
