% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_dataset_block}
\alias{new_dataset_block}
\title{Data block constructor}
\usage{
new_dataset_block(selected = character(), package = "datasets", ...)
}
\arguments{
\item{selected}{Selected dataset}

\item{package}{Name of an R package containing datasets}

\item{...}{Further (metadata) attributes}
}
\description{
This block allows to selected data from a package,
the default being datasets.
}
