% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_filesbrowser_block}
\alias{new_filesbrowser_block}
\title{Files browser block constructor}
\usage{
new_filesbrowser_block(volumes = c(home = path.expand("~")), ...)
}
\arguments{
\item{volumes}{Paths accessible by the shinyFiles browser.}

\item{...}{Further (metadata) attributes}
}
\description{
This block allows browse files on the server where the app is running.
It falls back to the user file system when running locally.
This block outputs a string containing the file path.
}
