% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_group_by_block}
\alias{new_group_by_block}
\title{Group by block}
\usage{
new_group_by_block(data, columns = character(), ...)
}
\arguments{
\item{data}{Tabular data in which to select some columns.}

\item{columns}{Column(s) to select.}

\item{...}{Further (metadata) attributes}
}
\description{
This block provides access to \link[dplyr]{group_by} verb and
returns a dataframe.
}
