% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_select_block}
\alias{new_select_block}
\title{Select block}
\usage{
new_select_block(columns = character(), ...)
}
\arguments{
\item{columns}{Column(s) to select.}

\item{...}{Further (metadata) attributes}
}
\description{
This block provides access to \link[dplyr]{select} verb and
returns a dataframe with the selected columns.
}
