% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{new_stack}
\alias{new_stack}
\alias{is_stack}
\alias{generate_code.stack}
\alias{block_combiner}
\alias{block_combiner.transform_block}
\alias{block_combiner.data_block}
\alias{block_combiner.plot_block}
\alias{block_combiner.plot_layer_block}
\alias{serve_stack}
\title{Stacks}
\usage{
new_stack(..., title = "Stack")

is_stack(x)

\method{generate_code}{stack}(x)

block_combiner(x, ...)

\method{block_combiner}{transform_block}(x, y, ...)

\method{block_combiner}{data_block}(x, y, ...)

\method{block_combiner}{plot_block}(x, y, ...)

\method{block_combiner}{plot_layer_block}(x, y, ...)

serve_stack(stack)
}
\arguments{
\item{...}{An ordered set of blocks (each argument is required to inherit
from \code{"block"})}

\item{title}{Stack title}

\item{x}{An object inheriting form \code{"stack"}}

\item{stack}{An object inheriting form \code{"stack"}}
}
\description{
A set of blocks can be assembled into a stack.
}
