% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_summarize_block}
\alias{new_summarize_block}
\title{Summarize block}
\usage{
new_summarize_block(
  func = character(),
  default_columns = character(),
  submit = NA,
  ...
)
}
\arguments{
\item{func}{Summarize function to apply.}

\item{default_columns}{If you know in advance each function to apply,
you can also pass predefined selected column for each summary.
Therefore when not of length 0, columns should have the same length
as func.}

\item{submit}{Whether this block requires to press a submit button to get
the results. Boolean, default to FALSE, which means no submit button.
If NA, then no computation is triggered and the user needs to click on the button
to see the block result. If TRUE, computation is automatically triggered and the button
is shown (which is useful when restoring a stack).}

\item{...}{Further (metadata) attributes}
}
\description{
This block provides access to \link[dplyr]{summarize} verb and
returns a dataframe with the transformed columns.
}
