% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_numeric_field}
\alias{new_numeric_field}
\title{Numeric field constructor}
\usage{
new_numeric_field(value = numeric(), min = numeric(), max = numeric(), ...)
}
\arguments{
\item{value}{Default text input value.}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{...}{Other parameters passed to \link{new_field} and may be needed
by \link{ui_input} to pass more options to the related shiny input.}
}
\description{
Numeric fields are translated into \link[shiny]{numericInput}
}
