% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_field2.R
\name{generate_server.numeric_field2}
\alias{generate_server.numeric_field2}
\alias{ui_input.numeric_field2}
\alias{new_numeric_field2}
\alias{numeric_field2}
\alias{validate_field.numeric_field2}
\title{Re-implement numeric_field, using server module}
\usage{
\method{generate_server}{numeric_field2}(x)

\method{ui_input}{numeric_field2}(x, id, name)

new_numeric_field2(value = numeric(), min = numeric(), max = numeric(), ...)

numeric_field2(...)

\method{validate_field}{numeric_field2}(x)
}
\description{
Re-implement numeric_field, using server module
}
