% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{off_canvas}
\alias{off_canvas}
\title{Bootstrap 5 offcanvas}
\usage{
off_canvas(id, title, ..., position = c("start", "top", "bottom", "end"))
}
\arguments{
\item{id}{Unique id. Must be triggered by a button
whose \code{data-bs-target} attributes matches this id.}

\item{title}{Title.}

\item{...}{Body content.}

\item{position}{Either \code{start} (left), \code{top}, \code{bottom}
or \code{end} (right).}
}
\value{
Boolean. TRUE if dependency found.
}
\description{
Sidebar like element either a top, bottom, right or left.
}
