% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_range_field}
\alias{new_range_field}
\alias{range_field}
\title{Range field constructor}
\usage{
new_range_field(value = numeric(), min = numeric(), max = numeric(), ...)

range_field(...)
}
\arguments{
\item{value}{Field value}

\item{min, max}{Slider boundaries (inclusive)}

\item{...}{Further field components}
}
\description{
Range fields are translated into \link[shiny]{sliderInput}.
}
