% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{remove_button}
\alias{remove_button}
\alias{remove_button.block}
\alias{remove_button.stack}
\title{Block remove generic}
\usage{
remove_button(x, ...)

\method{remove_button}{block}(x, id, ...)

\method{remove_button}{stack}(x, id, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}
}
\description{
Generic for creating block remove button.
}
