% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{remove_shiny_inputs}
\alias{remove_shiny_inputs}
\title{Cleanup module inputs}
\usage{
remove_shiny_inputs(.input, session = getDefaultReactiveDomain())
}
\arguments{
\item{.input}{Shiny input object.}

\item{session}{Shiny session object.}
}
\description{
Cleanup module inputs
}
\keyword{internal}
