% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{remove_stack_server}
\alias{remove_stack_server}
\title{Helper function that removes a stack}
\usage{
remove_stack_server(vals, input, session, id, workspace)
}
\description{
Helper function that removes a stack
}
