% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_result_block}
\alias{new_result_block}
\alias{result_block}
\title{Result block}
\usage{
new_result_block(...)

result_block(...)
}
\arguments{
\item{...}{Further (metadata) attributes}
}
\description{
A result blocks allows one to reuse the data from one \link{stack} into
another one. This isn't relevant for single stack apps.
}
