% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_result_field}
\alias{new_result_field}
\alias{result_field}
\title{Result field constructor}
\usage{
new_result_field(value = list(), ...)

result_field(...)
}
\arguments{
\item{value}{Field value}

\item{...}{Further field components}
}
\description{
Result field allows on to reuse the result of one stack
in another stack.
}
