% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secure}
\alias{secure}
\title{Evaluate expression safely}
\usage{
secure(expr, is_valid)
}
\arguments{
\item{expr}{Expression to evaluate.}

\item{is_valid}{Block valid status.}
}
\value{
Result or error message.
}
\description{
tryCatch wrapper.
}
\keyword{internal}
