% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_select_block}
\alias{new_select_block}
\alias{select_block}
\title{Select block}
\usage{
new_select_block(data, columns = colnames(data)[1], ...)

select_block(data, ...)
}
\arguments{
\item{data}{Tabular data in which to select some columns.}

\item{columns}{Column(s) to select.}

\item{...}{Further (metadata) attributes}
}
\description{
This block provides access to \link[dplyr]{select} verb and
returns a dataframe with the selected columns.
}
