% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_select_field}
\alias{new_select_field}
\alias{select_field}
\title{Select field constructor}
\usage{
new_select_field(
  value = character(),
  choices = character(),
  multiple = FALSE,
  ...
)

select_field(...)
}
\arguments{
\item{value}{Default text input value.}

\item{choices}{Select choices}

\item{multiple}{Allow multiple selection}

\item{...}{Other parameters passed to \link{new_field} and may be needed
by \link{ui_input} to pass more options to the related shiny input.}
}
\description{
Select fields are translated into \link[shiny]{selectInput}
}
