% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{send_error_to_ui}
\alias{send_error_to_ui}
\title{Sends validation error to user interface}
\usage{
send_error_to_ui(blk, is_valid, session)
}
\arguments{
\item{blk}{Block.}

\item{is_valid}{Block valid status.}

\item{session}{Shiny session object.}
}
\value{
Side effects.
}
\description{
Depending on whether some inputs are invalid.
}
\keyword{internal}
