% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialise.R
\name{serialise}
\alias{serialise}
\alias{serialise.field}
\alias{serialise.block}
\alias{serialise.stack}
\title{Serialise
Serialise stacks and blocks.}
\usage{
serialise(x, ...)

\method{serialise}{field}(x, ...)

\method{serialise}{block}(x, ...)

\method{serialise}{stack}(x, ...)
}
\arguments{
\item{x}{Object to serialise.}

\item{...}{Ignored.}
}
\description{
Serialise
Serialise stacks and blocks.
}
\section{Methods (by class)}{
\itemize{
\item \code{serialise(field)}: Serialise a field

\item \code{serialise(block)}: Serialise a block

\item \code{serialise(stack)}: Serialise a block

}}
