% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_code}
\alias{server_code}
\alias{server_code.block}
\title{Server code generic}
\usage{
server_code(x, state, output)

\method{server_code}{block}(x, state, output)
}
\arguments{
\item{x}{Block.}

\item{state}{Block state}

\item{output}{Shiny output}
}
\description{
Generate the server code output.
}
