% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_output}
\alias{server_output}
\alias{server_output.block}
\alias{server_output.upload_block}
\alias{server_output.filesbrowser_block}
\alias{server_output.plot_block}
\title{Server output generic}
\usage{
server_output(x, result, output)

\method{server_output}{block}(x, result, output)

\method{server_output}{upload_block}(x, result, output)

\method{server_output}{filesbrowser_block}(x, result, output)

\method{server_output}{plot_block}(x, result, output)
}
\arguments{
\item{x}{Block.}

\item{result}{Block result}

\item{output}{Shiny output}
}
\description{
Generate block server output. Needed on the UI
side. This is generally a table containing the
processed block data.
}
