% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{set_field_values_from_input}
\alias{set_field_values_from_input}
\title{Set field values of an object from a named input list}
\usage{
set_field_values_from_input(x, input)
}
\arguments{
\item{x}{An object with named fields that you want to update.}

\item{input}{A named list containing new values for the fields in \code{x}.}
}
\value{
An object with updated field values.
}
\description{
This function sets the fields of an object \code{x} using values from a named input list \code{input}.
It assumes that the field names in \code{x} and the names in the input list \code{input} match.
}
\examples{
\dontrun{
  x <- new_filter_block(dat = iris)
  input <- list(column = "Species", value = "versicolor")
  set_field_values_from_input(x, input)
}
}
