% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{set_workspace}
\alias{set_workspace}
\alias{add_workpace_stack}
\alias{rm_workspace_stack}
\alias{set_workspace_title}
\alias{set_workspace_settings}
\alias{list_workspace_stacks}
\alias{get_workspace}
\alias{get_workspace_stack}
\alias{get_workspace_stacks}
\alias{get_workspace_title}
\alias{get_workspace_settings}
\alias{serve_worskpace}
\title{Workspace}
\usage{
set_workspace(..., title = "", settings = NULL, force = FALSE)

add_workpace_stack(name, stack, force = FALSE)

rm_workspace_stack(name, force = FALSE)

set_workspace_title(title)

set_workspace_settings(settings)

list_workspace_stacks()

get_workspace()

get_workspace_stack(name)

get_workspace_stacks(names = list_workspace_stacks())

get_workspace_title()

get_workspace_settings()

serve_worskpace(clear = TRUE, id = "myworkspace")
}
\arguments{
\item{...}{A set of of stacks}

\item{title}{The workspace title}

\item{settings}{An (optional) list of settings (a string ist parsed as JSON)}

\item{force}{Force overwrite existing stacks (otherwise may throw warning)}

\item{name}{Stack name}

\item{stack}{A single stack}

\item{names}{Stack names}

\item{clear}{Clear current worspace}

\item{id}{Workspace ID}
}
\description{
Stacks all live in a singleton workspace.
}
