% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{stack_header}
\alias{stack_header}
\alias{stack_header.stack}
\title{Stack header generic}
\usage{
stack_header(x, ...)

\method{stack_header}{stack}(x, title, ns, add_block_func = add_block_ui, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{title}{Stack title.}

\item{ns}{Module namespace}

\item{add_block_func}{Function to add blocks within a stack. This allows
for customization. Default to \link{add_block_ui}.}
}
\description{
Generic for creating stack header.
}
