% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_string_field}
\alias{new_string_field}
\alias{string_field}
\title{String field constructor}
\usage{
new_string_field(value = character(), ...)

string_field(...)
}
\arguments{
\item{value}{Default text input value.}

\item{...}{Other parameters passed to \link{new_field} and may be needed
by \link{ui_input} to pass more options to the related shiny input.}
}
\description{
String fields are translated into \link[shiny]{textInput}
}
