% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_submit_field}
\alias{new_submit_field}
\alias{submit_field}
\title{Submit field constructor}
\usage{
new_submit_field(...)

submit_field(...)
}
\arguments{
\item{...}{Further field components}
}
\description{
Submit fields are translated into \link[shiny]{actionButton}
}
