% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_switch_field}
\alias{new_switch_field}
\alias{switch_field}
\title{Switch field constructor}
\usage{
new_switch_field(value = FALSE, ...)

switch_field(...)
}
\arguments{
\item{value}{Field value}

\item{...}{Further field components}
}
\description{
Switch fields are translated into \link[bslib]{input_switch}
}
