% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{title}
\alias{title}
\alias{set_title}
\alias{get_title}
\title{Title
Ge and set title.}
\usage{
set_title(stack, title)

get_title(stack)
}
\arguments{
\item{stack}{A stack.}

\item{title}{Title to set.}
}
\description{
Title
Ge and set title.
}
\keyword{internal}
