% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{to_json}
\alias{to_json}
\alias{to_json.field}
\alias{to_json.block}
\alias{to_json.stack}
\alias{to_json.workspace}
\alias{to_json.NULL}
\alias{from_json}
\title{JSON serialization}
\usage{
to_json(x = NULL)

\method{to_json}{field}(x)

\method{to_json}{block}(x)

\method{to_json}{stack}(x)

\method{to_json}{workspace}(x)

\method{to_json}{`NULL`}(x)

from_json(x)
}
\arguments{
\item{x}{Object to (de)serialize}
}
\description{
Object serialization as JSON.
}
