% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_fields}
\alias{ui_fields}
\alias{ui_fields.block}
\title{Block fields generic}
\usage{
ui_fields(x, ...)

\method{ui_fields}{block}(x, ns, inputs_hidden, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{inputs_hidden}{For styling purposes: CSS class to apply
when the block is collapsed.}
}
\description{
Generic for creating fields UI container
}
