% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{uiCode}
\alias{uiCode}
\alias{uiCode.block}
\title{Copy code generic}
\usage{
uiCode(x, ns)

\method{uiCode}{block}(x, ns)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{ns}{Module namespace}
}
\description{
Generate ui to copy block code and output
the code content.
}
