% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{uiOutputBlock}
\alias{uiOutputBlock}
\alias{uiOutputBlock.block}
\alias{uiOutputBlock.upload_block}
\alias{uiOutputBlock.filesbrowser_block}
\alias{uiOutputBlock.plot_block}
\title{Render block output generic}
\usage{
uiOutputBlock(x, ns)

\method{uiOutputBlock}{block}(x, ns)

\method{uiOutputBlock}{upload_block}(x, ns)

\method{uiOutputBlock}{filesbrowser_block}(x, ns)

\method{uiOutputBlock}{plot_block}(x, ns)
}
\arguments{
\item{x}{Block.}

\item{ns}{Output namespace}
}
\description{
Renders the block output.
}
