% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-core.R
\name{update_field}
\alias{update_field}
\alias{update_field.field}
\alias{update_field.hidden_field}
\alias{get_field_name}
\title{Update field generic}
\usage{
update_field(x, new, env = list())

\method{update_field}{field}(x, new, env = list())

\method{update_field}{hidden_field}(x, new, env = list())

get_field_name(field, name = "")
}
\arguments{
\item{x}{An object inheriting form \code{"field"}}

\item{new}{Value to set}

\item{env}{Environment with data and other field values}

\item{field}{Field element.}

\item{name}{To pass a name if no title attribute.}
}
\value{
The modified field.
}
\description{
Update a field with a new value.
Needed by \link{update_fields} in a block. Necessary to keep
the R object in sync with the Shiny interface state (input state).
}
