% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-core.R
\name{update_fields}
\alias{update_fields}
\alias{update_fields.data_block}
\alias{update_fields.transform_block}
\alias{update_fields.plot_block}
\title{Update fields generic}
\usage{
update_fields(x, ...)

\method{update_fields}{data_block}(x, session, ...)

\method{update_fields}{transform_block}(x, session, data, ...)

\method{update_fields}{plot_block}(x, session, data, ...)
}
\arguments{
\item{x}{An object inheriting from \code{"block"}}

\item{...}{For generic consistency.}

\item{session}{Shiny session}

\item{data}{Block input data}
}
\description{
For a block, update all its fields.
}
\seealso{
See \link{update_field}.
}
