% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_upload_block}
\alias{new_upload_block}
\alias{upload_block}
\title{Upload block constructor}
\usage{
new_upload_block(...)

upload_block(...)
}
\arguments{
\item{...}{Further (metadata) attributes}
}
\description{
This block allows to upload data from the user files system.
This block outputs a string containing the file path.
}
