% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_upload_field}
\alias{new_upload_field}
\title{Upload field constructor}
\usage{
new_upload_field(value = character(), ...)
}
\arguments{
\item{value}{Default text input value.}

\item{...}{Other parameters passed to \link{new_field} and may be needed
by \link{ui_input} to pass more options to the related shiny input.}
}
\description{
Upload fields are translated into \link[shiny]{fileInput}
}
