% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_block}
\alias{validate_block}
\title{Validate a entire block}
\usage{
validate_block(blk, is_valid, session)
}
\arguments{
\item{blk}{Block reactive value.}

\item{is_valid}{Block valid status.}

\item{session}{Shiny session object.}
}
\value{
Side effects.
}
\description{
Depending on whether some inputs are invalid.
}
\keyword{internal}
