% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-core.R, R/fields.R, R/keyvalue_field.R
\name{validate_field}
\alias{validate_field}
\alias{validate_field.string_field}
\alias{validate_field.select_field}
\alias{validate_field.switch_field}
\alias{validate_field.numeric_field}
\alias{validate_field.submit_field}
\alias{validate_field.upload_field}
\alias{validate_field.filesbrowser_field}
\alias{validate_field.variable_field}
\alias{validate_field.range_field}
\alias{validate_field.hidden_field}
\alias{validate_field.list_field}
\alias{validate_field.result_field}
\alias{validate_field.keyvalue_field}
\title{Validate field generic}
\usage{
validate_field(x)

\method{validate_field}{string_field}(x)

\method{validate_field}{select_field}(x)

\method{validate_field}{switch_field}(x)

\method{validate_field}{numeric_field}(x)

\method{validate_field}{submit_field}(x)

\method{validate_field}{upload_field}(x)

\method{validate_field}{filesbrowser_field}(x)

\method{validate_field}{variable_field}(x)

\method{validate_field}{range_field}(x)

\method{validate_field}{hidden_field}(x)

\method{validate_field}{list_field}(x)

\method{validate_field}{result_field}(x)

\method{validate_field}{keyvalue_field}(x)
}
\arguments{
\item{x}{An object inheriting form \code{"field"}}
}
\description{
Checks the value of a field with \link{value} and
apply corrections whenever necessary.
}
