% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate inputs}
\usage{
validate_inputs(blk, is_valid, session)
}
\arguments{
\item{blk}{Block reactive value.}

\item{is_valid}{Block valid status.}

\item{session}{Shiny session object.}
}
\value{
Side effects.
}
\description{
Get input value and determine if they're valid.
}
\keyword{internal}
