% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-core.R
\name{value}
\alias{value}
\alias{value.field}
\alias{value.variable_field}
\alias{value.list_field}
\alias{value<-}
\alias{value<-.field}
\alias{value<-.upload_field}
\alias{value<-.filesbrowser_field}
\title{Get field attribute value}
\usage{
value(x, name = "value")

\method{value}{field}(x, name = "value")

\method{value}{variable_field}(x, name = "value")

\method{value}{list_field}(x, name = "value")

value(x, name = "value") <- value

\method{value}{field}(x, name = "value") <- value

\method{value}{upload_field}(x, name = "value") <- value

\method{value}{filesbrowser_field}(x, name = "value") <- value
}
\arguments{
\item{x}{An object inheriting form \code{"field"}}

\item{name}{Field component name}

\item{value}{Field value}
}
\value{
Field value

The field.
}
\description{
Get the field value attribute. If it is a function,
return the result attribute instead.
}
