% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-core.R
\name{values}
\alias{values}
\title{Get all values from a field}
\usage{
values(x, name = names(x))
}
\arguments{
\item{x}{An object inheriting form \code{"field"}}

\item{name}{Field component name}
}
\value{
A list containing all values.
}
\description{
This calls \link{value} on all the field's names.
}
