% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R, R/stack.R, R/ui.R
\name{add_block_server}
\alias{add_block_server}
\alias{add_block_server.default}
\alias{add_block}
\alias{add_block_ui}
\alias{add_block_ui.default}
\title{Add block server generic}
\usage{
add_block_server(x, ...)

\method{add_block_server}{default}(x, id, vals, ...)

add_block(stack, block, position = NULL)

add_block_ui(x, ...)

\method{add_block_ui}{default}(x, id, ...)
}
\arguments{
\item{x}{Stack object.}

\item{...}{For generic consistency}

\item{id}{Module id.}

\item{vals}{Reactive values.}

\item{stack}{stack to update. See \link{new_stack}.}

\item{block}{Block to insert.}

\item{position}{Where to insert the new block. If NULL,
the block will be added in the last position.}
}
\value{
Invisibly returns the stack.
}
\description{
This modules aims at showing extra info in the
offcanvas menu to add blocks. Blocks are added
at the stack level with another function, add_block_stack.

This modules aims at showing extra info in the
offcanvas menu to add blocks. Blocks are added
at the stack level with another function, add_block_stack.

This is to be called oustide the stack by
other modules.

Useful to allow stack to add blocks to it.

Useful to allow stack to add blocks to it.
The selected block can be accessed through \code{input$search}
within \link{add_block_server}.
}
