% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{add_block_ui}
\alias{add_block_ui}
\title{Add block UI interface}
\usage{
add_block_ui(ns)
}
\arguments{
\item{ns}{Stack namespace.}
}
\description{
Useful to allow stack to add blocks to it.
The selected block can be accessed through \code{input$selected_block}.
Combined to the blocks registry API, this allows to select a block from R like
\code{available_blocks()[[input$selected_block]]}.
}
