% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{add_block_ui}
\alias{add_block_ui}
\title{Add block UI interface}
\usage{
add_block_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Useful to allow stack to add blocks to it.
The selected block can be accessed through \code{input$search}
within \link{add_block_server}.
}
