% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{add_block_ui}
\alias{add_block_ui}
\title{Add block UI interface}
\usage{
add_block_ui(x, ns = identity)
}
\arguments{
\item{x}{The stack object.}

\item{ns}{Stack namespace. Default to \link{identity} so
that it can be used when the stack is the top level element.}
}
\description{
Useful to allow stack to add blocks to it.
The selected block can be accessed through \code{input$selected_block}.
Combined to the blocks registry API, this allows to select a block from R
like \code{available_blocks()[[input$selected_block]]}.
}
