% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{available_blocks}
\alias{available_blocks}
\alias{block_name}
\alias{block_descr}
\title{Block registry}
\usage{
available_blocks()

block_name(x)

block_descr(x)
}
\arguments{
\item{x}{Block descriptino object}
}
\description{
List available blocks.
}
