% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{available_blocks}
\alias{available_blocks}
\alias{block_name}
\alias{block_descr}
\alias{register_block}
\alias{unregister_blocks}
\title{Block registry}
\usage{
available_blocks()

block_name(x)

block_descr(x)

register_block(
  constructor,
  name,
  description,
  classes,
  input,
  output,
  package = NA_character_
)

unregister_blocks(ids = NULL, package = NULL)
}
\arguments{
\item{x}{Block descriptino object}

\item{constructor}{Block constructor}

\item{name, description}{Metadata describing the block}

\item{classes}{Block classes}

\item{input, output}{Object types the block consumes and produces}

\item{package}{Package where block is defined}

\item{ids}{Character vector of block IDs (first entry in class attribute)}
}
\description{
List available blocks.
}
