% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{block_combiner}
\alias{block_combiner}
\alias{block_combiner.transform_block}
\alias{block_combiner.plot_block}
\alias{block_combiner.plot_layer_block}
\title{Combine 2 block expressions}
\usage{
block_combiner(left, right, ...)

\method{block_combiner}{transform_block}(left, right, ...)

\method{block_combiner}{plot_block}(left, right, ...)

\method{block_combiner}{plot_layer_block}(left, right, ...)
}
\arguments{
\item{left}{Left block object in \code{x \%>\% y}.}

\item{right}{Right block object in \code{x \%>\% y}.}

\item{...}{For generic consistency.}
}
\description{
Useful for \link{generate_code}.
}
