% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr-package}
\alias{blockr}
\alias{blockr-package}
\title{blockr: A block-based framework for data manipulation and visualization}
\description{
A framework for data manipulation and visualization using a web-based point and click user interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bristolmyerssquibb.github.io/blockr}
}

}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

Authors:
\itemize{
  \item David Granjon \email{david@cynkra.com}
  \item John Coene \email{john@opifex.org}
}

Other contributors:
\itemize{
  \item  Bristol Myers Squibb [copyright holder, funder]
}

}
\keyword{internal}
