% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{cleanup_block}
\alias{cleanup_block}
\title{Cleanup module inputs, outputs and observers}
\usage{
cleanup_block(input, output, obs)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{obs}{Module observer list. All modules observers
must be registered in a list containing them.}
}
\description{
Cleanup module inputs, outputs and observers
}
\keyword{internal}
