% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvalue_field.R
\name{exprs_ui}
\alias{exprs_ui}
\title{Create a UI element for expressions}
\usage{
exprs_ui(
  id = "",
  value_name = "newcol",
  value_val = NULL,
  delete_button = TRUE,
  key = c("suggest", "empty", "none")
)
}
\arguments{
\item{id}{Character string, an identifier for the UI element.}

\item{value_name}{Default name for the new column.}

\item{value_val}{Default value for the new column.}
}
\value{
A \code{div} element containing the UI components.
}
\description{
This function generates a UI element for inputting expressions in a Shiny application.
It includes two \code{shinyAce::aceEditor} elements for inputting the name and value of a new column.
}
\examples{
\dontrun{
library(shiny)
library(shinyAce)
shinyApp(
  ui = bslib::page_fluid(
    exprs_ui(value_name = "bla", value_val = "blabla")
  ),
  server = function(input, output) {}
)
}
}
