% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_filesbrowser_field}
\alias{new_filesbrowser_field}
\alias{filesbrowser_field}
\title{Files browser field constructor}
\usage{
new_filesbrowser_field(value = character(), ...)

filesbrowser_field(...)
}
\arguments{
\item{value}{Field value}

\item{...}{Further field components}
}
\description{
Files browser fields are translated into \link[shinyFiles]{shinyFilesButton}
}
