% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{generate_server}
\alias{generate_server}
\alias{generate_server.block}
\alias{generate_server.data_block}
\alias{generate_server.transform_block}
\alias{generate_server.stack}
\title{UI}
\usage{
generate_server(x, ...)

\method{generate_server}{block}(x, ...)

\method{generate_server}{data_block}(x, ...)

\method{generate_server}{transform_block}(x, in_dat, ...)

\method{generate_server}{stack}(x, ...)
}
\arguments{
\item{x}{Object for which to generate a \code{\link[shiny:moduleServer]{shiny::moduleServer()}}}

\item{...}{Generic consistency}

\item{in_dat}{Reactive input data}
}
\description{
Generic for UI generation
}
