% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{generate_server}
\alias{generate_server}
\alias{generate_server.block}
\alias{generate_server.data_block}
\alias{generate_server.transform_block}
\alias{generate_server.plot_block}
\alias{generate_server.ggiraph_block}
\alias{generate_server.stack}
\alias{handle_remove_block}
\alias{handle_remove_block.block}
\title{Server}
\usage{
generate_server(x, ...)

\method{generate_server}{block}(x, ...)

\method{generate_server}{data_block}(x, id, ...)

\method{generate_server}{transform_block}(x, in_dat, id, ...)

\method{generate_server}{plot_block}(x, in_dat, id, ...)

\method{generate_server}{ggiraph_block}(x, in_dat, id, ...)

\method{generate_server}{stack}(x, id = NULL, new_blocks = NULL, ...)

handle_remove_block(x, ...)

\method{handle_remove_block}{block}(x, vals, session = getDefaultReactiveDomain())
}
\arguments{
\item{x}{Block element.}

\item{...}{Generic consistency.}

\item{id}{Unique module id. Useful when the stack is called as a module.}

\item{in_dat}{Reactive input data}

\item{new_blocks}{For dynamically inserted blocks.}

\item{vals}{Internal stack reactive values.}

\item{session}{Shiny session object.}
}
\description{
Generic for server generation

Generic for block removal

Necessary to be able to remove the block from the stack.
}
