% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{generate_server}
\alias{generate_server}
\alias{generate_server.result_field}
\alias{generate_server.data_block}
\alias{generate_server.transform_block}
\alias{generate_server.plot_block}
\alias{generate_server.stack}
\alias{generate_server.workspace}
\title{Generic for server generation}
\usage{
generate_server(x, ...)

\method{generate_server}{result_field}(x, ...)

\method{generate_server}{data_block}(x, id, linked_stack, ...)

\method{generate_server}{transform_block}(x, in_dat, id, is_prev_valid, ...)

\method{generate_server}{plot_block}(x, in_dat, id, is_prev_valid, ...)

\method{generate_server}{stack}(
  x,
  id = NULL,
  new_block = NULL,
  workspace = get_workspace(),
  prev_stack,
  ...
)

\method{generate_server}{workspace}(x, id, ...)
}
\arguments{
\item{x}{Object for which to generate a \code{\link[=moduleServer]{moduleServer()}}}

\item{...}{Generic consistency}

\item{id}{Unique module id. Useful when the workspace is called as a module.}

\item{in_dat}{Reactive input data}

\item{is_prev_valid}{Useful to validate the current block}

\item{new_block}{For dynamically inserted blocks.}

\item{workspace}{Stack workspace}
}
\description{
Calls shiny modules for the given element (workspace, stack, block).
}
