% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvalue_field.R, R/server.R
\name{generate_server.keyvalue_field}
\alias{generate_server.keyvalue_field}
\alias{generate_server}
\alias{generate_server.block}
\alias{generate_server.data_block}
\alias{generate_server.upload_block}
\alias{generate_server.transform_block}
\alias{generate_server.plot_block}
\alias{generate_server.stack}
\alias{handle_remove}
\alias{handle_remove.block}
\alias{handle_remove.stack}
\alias{generate_server.workspace}
\alias{init}
\alias{init.workspace}
\alias{init.stack}
\title{Key Value Field}
\usage{
\method{generate_server}{keyvalue_field}(x, ...)

generate_server(x, ...)

\method{generate_server}{block}(x, ...)

\method{generate_server}{data_block}(x, id, ...)

\method{generate_server}{upload_block}(x, id, ...)

\method{generate_server}{transform_block}(x, in_dat, id, ...)

\method{generate_server}{plot_block}(x, in_dat, id, ...)

\method{generate_server}{stack}(x, id = NULL, new_block = NULL, ...)

handle_remove(x, ...)

\method{handle_remove}{block}(x, vals, session = getDefaultReactiveDomain(), ...)

\method{handle_remove}{stack}(x, vals, session = getDefaultReactiveDomain(), ...)

\method{generate_server}{workspace}(x, id = NULL, ...)

init(x, ...)

\method{init}{workspace}(x, stacks, vals, session, ...)

\method{init}{stack}(x, vals, session, ...)
}
\arguments{
\item{x}{Element.}

\item{...}{Generic consistency.}

\item{id}{Unique module id. Useful when the workspace is called as a module.}

\item{in_dat}{Reactive input data}

\item{new_block}{For dynamically inserted blocks.}

\item{vals}{Internal reactive values.}

\item{session}{Shiny session object.}

\item{stacks}{List of workspace stacks.}
}
\description{
Generic for server generation

Generic for stack/block removal

Necessary to be able to remove the block from the stack.

Necessary to be able to remove the stack from the workspace.

Handle initialisation of workspace, stacks, ...
}
