% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{generate_ui}
\alias{generate_ui}
\alias{generate_ui.block}
\alias{generate_ui.stack}
\alias{ui_input}
\alias{ui_input.field}
\alias{ui_input.string_field}
\alias{ui_input.select_field}
\alias{ui_update}
\alias{ui_update.field}
\alias{ui_update.string_field}
\alias{ui_update.select_field}
\title{UI}
\usage{
generate_ui(x, ...)

\method{generate_ui}{block}(x, id, ...)

\method{generate_ui}{stack}(x, ...)

ui_input(x, id, name)

\method{ui_input}{field}(x, id, name)

\method{ui_input}{string_field}(x, id, name)

\method{ui_input}{select_field}(x, id, name)

ui_update(x, session, id, name)

\method{ui_update}{field}(x, session, id, name)

\method{ui_update}{string_field}(x, session, id, name)

\method{ui_update}{select_field}(x, session, id, name)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{id}{UI IDs}

\item{name}{Field name}

\item{session}{Shiny session}
}
\description{
Generic for UI generation
}
