% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R, R/ui.R
\name{server_output}
\alias{server_output}
\alias{server_output.block}
\alias{server_code}
\alias{server_code.block}
\alias{generate_ui}
\alias{generate_ui.block}
\alias{generate_ui.stack}
\alias{ui_input}
\alias{ui_input.field}
\alias{ui_input.string_field}
\alias{ui_input.select_field}
\alias{ui_output}
\alias{ui_output.plot_block}
\alias{ui_update}
\alias{ui_update.field}
\alias{ui_update.string_field}
\alias{ui_update.select_field}
\alias{ui_output.block}
\alias{ui_code}
\alias{ui_code.block}
\title{UI}
\usage{
server_output(x, result, output)

\method{server_output}{block}(x, result, output)

server_code(x, state, output)

\method{server_code}{block}(x, state, output)

generate_ui(x, ...)

\method{generate_ui}{block}(x, id, ...)

\method{generate_ui}{stack}(x, ...)

ui_input(x, id, name)

\method{ui_input}{field}(x, id, name)

\method{ui_input}{string_field}(x, id, name)

\method{ui_input}{select_field}(x, id, name)

ui_output(x, ns)

\method{ui_output}{plot_block}(x, id)

ui_update(x, session, id, name)

\method{ui_update}{field}(x, session, id, name)

\method{ui_update}{string_field}(x, session, id, name)

\method{ui_update}{select_field}(x, session, id, name)

ui_output(x, ns)

\method{ui_output}{block}(x, ns)

ui_code(x, ns)

\method{ui_code}{block}(x, ns)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{result}{Block result}

\item{output}{Shiny output}

\item{state}{Block state}

\item{...}{Generic consistency}

\item{id}{UI IDs}

\item{name}{Field name}

\item{ns}{Output namespace}

\item{session}{Shiny session}
}
\description{
Generic for UI generation
}
