% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{generate_ui}
\alias{generate_ui}
\alias{generate_ui.block}
\alias{generate_ui.stack}
\alias{generate_ui.workspace}
\alias{ui_update.expression_field}
\title{UI}
\usage{
generate_ui(x, ...)

\method{generate_ui}{block}(x, id, ..., .hidden = !getOption("BLOCKR_DEV", FALSE))

\method{generate_ui}{stack}(x, id = NULL, ...)

\method{generate_ui}{workspace}(x, id, ...)

\method{ui_update}{expression_field}(x, session, id, name)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{id}{UI IDs}

\item{.hidden}{Whether to initialise the block hidden.}
}
\description{
Generic for UI generation
}
