% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-core.R, R/field-core.R, R/fields.R
\name{is_initialized}
\alias{is_initialized}
\alias{is_initialized.block}
\alias{is_initialized.field}
\alias{is_initialized.variable_field}
\alias{is_initialized.list_field}
\title{S3 generic for initialization}
\usage{
is_initialized(x)

\method{is_initialized}{block}(x)

\method{is_initialized}{field}(x)

\method{is_initialized}{variable_field}(x)

\method{is_initialized}{list_field}(x)
}
\arguments{
\item{x}{Element to check.}
}
\value{
Boolean value.
}
\description{
Checks if a block or field (currently implemented methods)
is initialized.
}
