% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{inject_remove_button}
\alias{inject_remove_button}
\title{Inject remove button into stack/block header}
\usage{
inject_remove_button(ns, el)
}
\arguments{
\item{ns}{Parent namespace.}

\item{el}{Current block or stack.}
}
\description{
This has to be called from the workspace/stack parent
namespace. This can also be called dynamically when
inserting a new stack/block within a workspace/stack.
}
\keyword{internal}
