% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{inject_remove_button}
\alias{inject_remove_button}
\alias{inject_remove_button.block}
\alias{inject_remove_button.stack}
\title{Inject remove button into block header}
\usage{
inject_remove_button(x, ...)

\method{inject_remove_button}{block}(x, ns, .hidden = !getOption("BLOCKR_DEV", FALSE), ...)

\method{inject_remove_button}{stack}(x, id, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{ns}{Parent namespace.}

\item{.hidden}{Whether to initialise the block with
hidden inputs.}

\item{id}{Parent ID}
}
\description{
This has to be called from the stack parent
namespace. This can also be called dynamically when
inserting a new block within a stack.

This has to be called from the workspace parent
namespace. This can also be called dynamically when
inserting a new stack within a workspace.
}
