% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvalue_field.R
\name{generate_server.keyvalue_field}
\alias{generate_server.keyvalue_field}
\alias{ui_input.keyvalue_field}
\alias{new_keyvalue_field}
\alias{keyvalue_field}
\title{Key Value Field server}
\usage{
\method{generate_server}{keyvalue_field}(x, ...)

\method{ui_input}{keyvalue_field}(x, id, name)

new_keyvalue_field(
  value = NULL,
  multiple = TRUE,
  submit = TRUE,
  key = c("suggest", "empty", "none"),
  ...
)

keyvalue_field(...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}

\item{id}{UI IDs}

\item{name}{Field name.}

\item{value}{Field value}

\item{multiple}{Allow multiple selection}

\item{submit}{Should a 'submit button' be shown?}

\item{key}{How to display the 'key' field}
}
\description{
Server module

TBD
}
