% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_list_field}
\alias{new_list_field}
\title{List field constructor}
\usage{
new_list_field(
  field = character(),
  components = list(),
  name = character(),
  package = "blockr",
  ...
)
}
\arguments{
\item{field}{Field type}

\item{components}{Variable list of field components}

\item{name}{Sub-field name(s)}

\item{package}{Package weher to find the field constructor (as
\verb{new_\{class_name\}})}

\item{...}{Other parameters passed to \link{new_field} and may be needed
by \link{ui_input} to pass more options to the related shiny input.}
}
\description{
A field that can contain subfields. See \link{new_filter_block} for
a usecase.
}
