% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{new_block}
\alias{new_block}
\alias{generate_code}
\alias{generate_code.block}
\alias{evalute_block}
\alias{evalute_block.block}
\alias{generate_ui}
\alias{generate_ui.block}
\alias{new_data_block}
\title{Blocks}
\usage{
new_block(name, fields, expr, ..., class = character())

generate_code(x)

\method{generate_code}{block}(x)

evalute_block(x)

\method{evalute_block}{block}(x)

generate_ui(x)

\method{generate_ui}{block}(x)

new_data_block()
}
\arguments{
\item{name}{Block name}

\item{fields}{A list of field, each entry inheriting from \code{"fields"}}

\item{expr}{A quoted expression (compatible with partial substitution as
implemented in \code{\link[base:bquote]{base::bquote()}} and intended for evaluation in the context
of the fields)}

\item{...}{Further (metadata) attributes}

\item{class}{Block subclass}

\item{x}{An object inheriting form \code{"block"}}
}
\description{
Blocks consist of a set of fields, an expression that defines what the block
produces (given the result of the previous block combined with user input),
plus some meta data.
}
