% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{new_block}
\alias{new_block}
\alias{is_block}
\alias{generate_code}
\alias{generate_code.block}
\alias{evalute_block}
\alias{evalute_block.block}
\alias{evalute_block.data_block}
\alias{evalute_block.transform_block}
\alias{evalute_block.plot_block}
\alias{new_data_block}
\alias{new_filter_block}
\alias{new_select_block}
\alias{new_plot_block}
\alias{update_fields}
\alias{update_fields.block}
\alias{update_fields.filter_block}
\title{Blocks}
\usage{
new_block(fields, expr, name = rand_names(), ..., class = character())

is_block(x)

generate_code(x)

\method{generate_code}{block}(x)

evalute_block(x, ...)

\method{evalute_block}{block}(x, ...)

\method{evalute_block}{data_block}(x, ...)

\method{evalute_block}{transform_block}(x, data, ...)

\method{evalute_block}{plot_block}(x, data, ...)

new_data_block(...)

new_filter_block(dat, col = colnames(dat)[1L], val = NA_character_, ...)

new_select_block(dat, cols = colnames(dat)[1L], ...)

new_plot_block(dat, x, y, plot_opts = list(color = "blue"), ...)

update_fields(x, ...)

\method{update_fields}{block}(x, ...)

\method{update_fields}{filter_block}(x, data, session, ...)
}
\arguments{
\item{fields}{A list of field, each entry inheriting from \code{"field"}}

\item{expr}{A quoted expression (compatible with partial substitution as
implemented in \code{\link[base:bquote]{base::bquote()}} and intended for evaluation in the context
of the fields)}

\item{name}{Block name}

\item{...}{Any other params. TO DO}

\item{class}{Block subclass}

\item{x}{X axis variable.}

\item{data}{Result from previous block}

\item{dat}{Tabular data in which to select some columns.}

\item{col, val}{Definition of the equality filter}

\item{cols}{Column(s) to select.}

\item{y}{Y axis variable.}

\item{plot_opts}{List containing options for ggplot (color, ...).}

\item{session}{Shiny session}
}
\description{
Blocks consist of a set of fields, an expression that defines what the block
produces (given the result of the previous block combined with user input),
plus some meta data.
}
