% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R, R/field.R
\name{new_block}
\alias{new_block}
\alias{is_block}
\alias{is_initialized}
\alias{is_initialized.block}
\alias{initialize_block}
\alias{generate_code}
\alias{generate_code.block}
\alias{generate_code.transform_block}
\alias{evalute_block}
\alias{evalute_block.data_block}
\alias{evalute_block.transform_block}
\alias{evalute_block.plot_block}
\alias{new_data_block}
\alias{data_block}
\alias{initialize_block.data_block}
\alias{new_filter_block}
\alias{filter_block}
\alias{new_select_block}
\alias{new_summarize_block}
\alias{summarize_block}
\alias{select_block}
\alias{arrange_block}
\alias{group_by_block}
\alias{new_plot_block}
\alias{plot_block}
\alias{initialize_block.transform_block}
\alias{initialize_block.plot_block}
\alias{update_fields}
\alias{update_fields.data_block}
\alias{update_fields.transform_block}
\alias{update_fields.plot_block}
\alias{is_initialized.field}
\title{Blocks}
\usage{
new_block(fields, expr, name = rand_names(), ..., class = character())

is_block(x)

is_initialized(x)

\method{is_initialized}{block}(x)

initialize_block(x, ...)

generate_code(x)

\method{generate_code}{block}(x)

\method{generate_code}{transform_block}(x)

evalute_block(x, ...)

\method{evalute_block}{data_block}(x, ...)

\method{evalute_block}{transform_block}(x, data, ...)

\method{evalute_block}{plot_block}(x, data, ...)

new_data_block(...)

data_block(...)

\method{initialize_block}{data_block}(x, ...)

new_filter_block(data, columns = "LBTEST", values = character(), ...)

filter_block(data, ...)

new_select_block(data, columns = colnames(data)[1], ...)

new_summarize_block(data, column = "LBSTRESN", func = c("mean", "sd"), ...)

summarize_block(data, ...)

select_block(data, ...)

arrange_block(data, ...)

group_by_block(data, ...)

new_plot_block(
  data,
  plot_opts = list(colors = c("blue", "red"), point_size = 3, title = "Plot title", theme
    = c("theme_minimal", "theme_gray", "theme_linedraw", "theme_dark", "theme_light",
    "theme_classic", "theme_void", "theme_bw"), x_lab = "X axis label", y_lab =
    "Y axis label"),
  ...
)

plot_block(data, ...)

\method{initialize_block}{transform_block}(x, data, ...)

\method{initialize_block}{plot_block}(x, data, ...)

update_fields(x, ...)

\method{update_fields}{data_block}(x, session, ...)

\method{update_fields}{transform_block}(x, session, data, ...)

\method{update_fields}{plot_block}(x, session, data, ...)

\method{is_initialized}{field}(x)
}
\arguments{
\item{fields}{A list of field, each entry inheriting from \code{"field"}}

\item{expr}{A quoted expression (compatible with partial substitution as
implemented in \code{\link[base:bquote]{base::bquote()}} and intended for evaluation in the context
of the fields)}

\item{name}{Block name}

\item{...}{Any other params. TO DO}

\item{class}{Block subclass}

\item{x}{An object inheriting form \code{"block"}}

\item{data}{Block input data}

\item{columns}{Column(s) to select.}

\item{values}{Definition of the equality filter.}

\item{column}{Column to select.}

\item{func}{Summarize function to apply.}

\item{plot_opts}{List containing options for ggplot (color, ...).}

\item{session}{Shiny session}
}
\description{
Blocks consist of a set of fields, an expression that defines what the block
produces (given the result of the previous block combined with user input),
plus some meta data.
}
