% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R, R/field.R, R/mutate-block.R
\name{new_block}
\alias{new_block}
\alias{is_block}
\alias{is_initialized}
\alias{is_initialized.block}
\alias{initialize_block}
\alias{generate_code}
\alias{generate_code.block}
\alias{generate_code.arrange_block}
\alias{generate_code.transform_block}
\alias{generate_code.data_block}
\alias{generate_code.plot_block}
\alias{generate_code.call}
\alias{evaluate_block}
\alias{evaluate_block.data_block}
\alias{evaluate_block.transform_block}
\alias{evaluate_block.plot_block}
\alias{evaluate_block.plot_layer_block}
\alias{new_data_block}
\alias{data_block}
\alias{new_upload_block}
\alias{upload_block}
\alias{new_filesbrowser_block}
\alias{filesbrowser_block}
\alias{new_parser_block}
\alias{new_csv_block}
\alias{csv_block}
\alias{new_rds_block}
\alias{rds_block}
\alias{new_json_block}
\alias{json_block}
\alias{new_sas_block}
\alias{sas_block}
\alias{new_xpt_block}
\alias{xpt_block}
\alias{new_result_block}
\alias{result_block}
\alias{initialize_block.data_block}
\alias{new_filter_block}
\alias{filter_block}
\alias{new_select_block}
\alias{new_summarize_block}
\alias{summarize_block}
\alias{select_block}
\alias{new_arrange_block}
\alias{arrange_block}
\alias{new_group_by_block}
\alias{group_by_block}
\alias{new_join_block}
\alias{join_block}
\alias{new_head_block}
\alias{head_block}
\alias{initialize_block.transform_block}
\alias{initialize_block.default}
\alias{initialize_block.plot_block}
\alias{update_fields}
\alias{update_fields.data_block}
\alias{update_fields.transform_block}
\alias{update_fields.plot_block}
\alias{is_initialized.field}
\alias{mutate_block}
\title{Blocks}
\usage{
new_block(fields, expr, name = rand_names(), ..., class = character())

is_block(x)

is_initialized(x)

\method{is_initialized}{block}(x)

initialize_block(x, ...)

generate_code(x)

\method{generate_code}{block}(x)

\method{generate_code}{arrange_block}(x)

\method{generate_code}{transform_block}(x)

\method{generate_code}{data_block}(x)

\method{generate_code}{plot_block}(x)

\method{generate_code}{call}(x)

evaluate_block(x, ...)

\method{evaluate_block}{data_block}(x, ...)

\method{evaluate_block}{transform_block}(x, data, ...)

\method{evaluate_block}{plot_block}(x, data, ...)

\method{evaluate_block}{plot_layer_block}(x, data, ...)

new_data_block(
  ...,
  dat = as.environment("package:datasets"),
  selected = character()
)

data_block(...)

new_upload_block(...)

upload_block(...)

new_filesbrowser_block(volumes = c(home = path.expand("~")), ...)

filesbrowser_block(...)

new_parser_block(data, expr, fields = list(), ..., class = character())

new_csv_block(data, ...)

csv_block(data, ...)

new_rds_block(data, ...)

rds_block(data, ...)

new_json_block(data, ...)

json_block(data, ...)

new_sas_block(data, ...)

sas_block(data, ...)

new_xpt_block(data, ...)

xpt_block(data, ...)

new_result_block(...)

result_block(...)

\method{initialize_block}{data_block}(x, ...)

new_filter_block(
  data,
  columns = colnames(data)[1L],
  values = character(),
  filter_fun = "==",
  ...
)

filter_block(data, ...)

new_select_block(data, columns = colnames(data)[1], ...)

new_summarize_block(
  data,
  func = c("mean", "se"),
  default_columns = character(),
  ...
)

summarize_block(data, ...)

select_block(data, ...)

new_arrange_block(data, columns = colnames(data)[1], ...)

arrange_block(data, ...)

new_group_by_block(data, columns = colnames(data)[1], ...)

group_by_block(data, ...)

new_join_block(data, y = NULL, type = character(), by = character(), ...)

join_block(data, ...)

new_head_block(data, n_rows = numeric(), n_rows_min = 1L, ...)

head_block(data, ...)

\method{initialize_block}{transform_block}(x, data, ...)

\method{initialize_block}{default}(x, data, ...)

\method{initialize_block}{plot_block}(x, data, ...)

update_fields(x, ...)

\method{update_fields}{data_block}(x, session, ...)

\method{update_fields}{transform_block}(x, session, data, ...)

\method{update_fields}{plot_block}(x, session, data, ...)

\method{is_initialized}{field}(x)

mutate_block(data, ...)
}
\arguments{
\item{fields}{A list of field, each entry inheriting from \code{"field"}}

\item{expr}{A quoted expression (compatible with partial substitution as
implemented in \code{\link[base:bquote]{base::bquote()}} and intended for evaluation in the context
of the fields)}

\item{name}{Block name}

\item{...}{Further (metadata) attributes}

\item{class}{Block subclass}

\item{x}{An object inheriting form \code{"block"}}

\item{data}{Block input data}

\item{dat}{Multiple datasets.}

\item{selected}{Selected dataset.}

\item{volumes}{Paths accessible by the shinyFiles browser.}

\item{columns}{Column(s) to select.}

\item{values}{Definition of the equality filter.}

\item{filter_fun}{Default filter fun for the expression.}

\item{func}{Summarize function to apply.}

\item{default_columns}{If you know in advance each function to apply,
you can also pass predefined selected column for each summary.
Therefore when not of length 0, columns should have the same length
as func.}

\item{y}{Second dataset for join}

\item{type}{Join type}

\item{by}{Join columns}

\item{n_rows}{Number of rows to return.}

\item{n_rows_min}{Minimum number of rows.}

\item{session}{Shiny session}
}
\description{
Blocks consist of a set of fields, an expression that defines what the block
produces (given the result of the previous block combined with user input),
plus some meta data.
}
\note{
Useful to build other parser blocks
}
\keyword{internal}
