% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-core.R, R/fields.R
\name{new_block}
\alias{new_block}
\alias{is_block}
\alias{is_initialized.expression_field}
\alias{is_initialized.kv_field}
\title{Blocks}
\usage{
new_block(fields, expr, name = rand_names(), ..., class = character())

is_block(x)

\method{is_initialized}{expression_field}(x)

\method{is_initialized}{kv_field}(x)
}
\arguments{
\item{fields}{A list of field, each entry inheriting from \code{"field"}}

\item{expr}{A quoted expression (compatible with partial substitution as
implemented in \code{\link[base:bquote]{base::bquote()}} and intended for evaluation in the context
of the fields)}

\item{name}{Block name}

\item{...}{Further (metadata) attributes}

\item{class}{Block subclass}

\item{x}{An object inheriting from \code{"block"}}
}
\description{
Blocks consist of a set of fields, an expression that defines what the block
produces (given the result of the previous block combined with user input),
plus some meta data.
}
