% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{new_field}
\alias{new_field}
\alias{validate_field}
\alias{validate_field.field}
\alias{is_field}
\alias{validate_field.string_field}
\alias{new_string_field}
\alias{string_field}
\alias{validate_field.select_field}
\alias{new_select_field}
\alias{select_field}
\alias{new_field_list}
\alias{is_field_list}
\title{Fields}
\usage{
new_field(value, ..., type = c("literal", "name"), class = character())

validate_field(x)

\method{validate_field}{field}(x)

is_field(x)

\method{validate_field}{string_field}(x)

new_string_field(value = character(), ...)

string_field(...)

\method{validate_field}{select_field}(x)

new_select_field(value = character(), choices = character(), ...)

select_field(...)

new_field_list(...)

is_field_list(x)
}
\arguments{
\item{value}{Field value}

\item{...}{Further (metadata) attributes}

\item{type}{Field type (allowed values are \code{"literal"} and \code{"name"})}

\item{class}{Field subclass}

\item{x}{An object inheriting form \code{"field"}}

\item{choices}{Set of permissible values}
}
\description{
Each block consists of a set of fields, which define the type of value
the field holds and can be used to customize how the UI is generated.
}
