% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{new_field}
\alias{new_field}
\alias{validate_field}
\alias{validate_field.field}
\alias{ui_input}
\alias{ui_input.field}
\alias{is_field}
\alias{validate_field.string_field}
\alias{string_field}
\alias{ui_input.string_field}
\alias{validate_field.select_field}
\alias{select_field}
\alias{ui_input.select_field}
\title{Fields}
\usage{
new_field(value, ..., class = character())

validate_field(x)

\method{validate_field}{field}(x)

ui_input(id, name, x)

\method{ui_input}{field}(id, name, x)

is_field(x)

\method{validate_field}{string_field}(x)

string_field(value)

\method{ui_input}{string_field}(id, name, x)

\method{validate_field}{select_field}(x)

select_field(value, choices)

\method{ui_input}{select_field}(id, name, x)
}
\arguments{
\item{value}{Field value}

\item{...}{Further (metadata) attributes}

\item{class}{Field subclass}

\item{x}{An object inheriting form \code{"field"}}

\item{id, name}{Field ID and name}

\item{choices}{Set of permissible values}
}
\description{
Each block consists of a set of fields, which define the type of value
the field holds and can be used to customize how the UI is generated.
}
