% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-core.R, R/fields.R
\name{new_field}
\alias{new_field}
\alias{is_field}
\alias{new_expression_field}
\alias{validate_field.expression_field}
\title{Fields}
\usage{
new_field(
  value,
  ...,
  type = c("literal", "name"),
  title = "",
  descr = "",
  status = c("active", "disabled", "invisible"),
  class = character()
)

is_field(x)

new_expression_field(value = "", autocomplete = character(), ...)

\method{validate_field}{expression_field}(x)
}
\arguments{
\item{value}{Field value}

\item{...}{Further field components}

\item{type}{Field type (allowed values are \code{"literal"} and \code{"name"})}

\item{title}{A brief title for the field, primarily for display purposes.}

\item{descr}{A description of the field, explaining its purpose or usage.}

\item{status}{The status of the field (experimental)}

\item{class}{Field subclass
feature. Default to FALSE. Not yet used.}

\item{x}{An object inheriting form \code{"field"}}

\item{autocomplete}{Additional autocomplete options}
}
\description{
Each block consists of a set of fields, which define the type of value
the field holds and can be used to customize how the UI is generated.
}
