% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R, R/keyvalue_field.R
\name{new_field}
\alias{new_field}
\alias{validate_field}
\alias{update_field}
\alias{update_field.field}
\alias{initialize_field}
\alias{initialize_field.field}
\alias{is_field}
\alias{validate_field.string_field}
\alias{new_string_field}
\alias{string_field}
\alias{new_select_field}
\alias{select_field}
\alias{validate_field.select_field}
\alias{new_switch_field}
\alias{switch_field}
\alias{validate_field.switch_field}
\alias{new_numeric_field}
\alias{numeric_field}
\alias{validate_field.numeric_field}
\alias{new_submit_field}
\alias{submit_field}
\alias{validate_field.submit_field}
\alias{new_upload_field}
\alias{upload_field}
\alias{validate_field.upload_field}
\alias{new_filesbrowser_field}
\alias{filesbrowser_field}
\alias{validate_field.filesbrowser_field}
\alias{value}
\alias{values}
\alias{value<-}
\alias{new_variable_field}
\alias{variable_field}
\alias{validate_field.variable_field}
\alias{new_range_field}
\alias{range_field}
\alias{validate_field.range_field}
\alias{new_hidden_field}
\alias{hidden_field}
\alias{validate_field.hidden_field}
\alias{new_list_field}
\alias{list_field}
\alias{validate_field.list_field}
\alias{new_result_field}
\alias{result_field}
\alias{validate_field.result_field}
\alias{new_keyvalue_field}
\alias{keyvalue_field}
\alias{validate_field.keyvalue_field}
\title{Fields}
\usage{
new_field(
  value,
  ...,
  type = c("literal", "name"),
  title = "",
  descr = "",
  status = c("active", "disabled", "invisible"),
  class = character(),
  exclude = FALSE
)

validate_field(x)

update_field(x, new, env = list())

\method{update_field}{field}(x, new, env = list())

initialize_field(x, env = list())

\method{initialize_field}{field}(x, env = list())

is_field(x)

\method{validate_field}{string_field}(x)

new_string_field(value = character(), ...)

string_field(...)

new_select_field(
  value = character(),
  choices = character(),
  multiple = FALSE,
  ...
)

select_field(...)

\method{validate_field}{select_field}(x)

new_switch_field(value = FALSE, ...)

switch_field(...)

\method{validate_field}{switch_field}(x)

new_numeric_field(value = numeric(), min = numeric(), max = numeric(), ...)

numeric_field(...)

\method{validate_field}{numeric_field}(x)

new_submit_field(...)

submit_field(...)

\method{validate_field}{submit_field}(x)

new_upload_field(value = character(), ...)

upload_field(...)

\method{validate_field}{upload_field}(x)

new_filesbrowser_field(
  value = character(),
  volumes = c(home = path.expand("~")),
  ...
)

filesbrowser_field(...)

\method{validate_field}{filesbrowser_field}(x)

value(x, name = "value")

values(x, name = names(x))

value(x, name = "value") <- value

new_variable_field(
  value = character(),
  field = character(),
  components = list(),
  ...
)

variable_field(...)

\method{validate_field}{variable_field}(x)

new_range_field(value = numeric(), min = numeric(), max = numeric(), ...)

range_field(...)

\method{validate_field}{range_field}(x)

new_hidden_field(value = expression(), ...)

hidden_field(...)

\method{validate_field}{hidden_field}(x)

new_list_field(value = list(), sub_fields = list(), ...)

list_field(...)

\method{validate_field}{list_field}(x)

new_result_field(value = list(), ...)

result_field(...)

\method{validate_field}{result_field}(x)

new_keyvalue_field(
  value = NULL,
  multiple = TRUE,
  submit = TRUE,
  key = c("suggest", "empty", "none"),
  ...
)

keyvalue_field(...)

\method{validate_field}{keyvalue_field}(x)
}
\arguments{
\item{value}{Field value}

\item{...}{Further field components}

\item{type}{Field type (allowed values are \code{"literal"} and \code{"name"})}

\item{title}{A brief title for the field, primarily for display purposes.}

\item{descr}{A description of the field, explaining its purpose or usage.}

\item{status}{The status of the field (experimental)}

\item{class}{Field subclass}

\item{exclude}{Experimental: Exclude field from being captured in the update_fields
feature. Default to FALSE. Not yet used.}

\item{x}{An object inheriting form \code{"field"}}

\item{new}{Value to set}

\item{env}{Environment with data and other field values}

\item{choices}{Set of permissible values}

\item{multiple}{Allow multiple selections}

\item{min, max}{Slider boundaries (inclusive)}

\item{volumes}{Paths accessible by the shinyFiles browser.}

\item{name}{Field component name}

\item{field}{Field type}

\item{components}{Variable list of field components}

\item{sub_fields}{Fields contained in \code{list_field}}

\item{submit}{Should a 'submit button' be shown?}

\item{key}{How to display the 'key' field}
}
\description{
Each block consists of a set of fields, which define the type of value
the field holds and can be used to customize how the UI is generated.
}
