% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{new_filter_block}
\alias{new_filter_block}
\title{Filter block}
\usage{
new_filter_block(
  columns = character(),
  values = character(),
  filter_fun = "==",
  ...
)
}
\arguments{
\item{columns}{Columns used for filtering}

\item{values}{Values used for filtering}

\item{filter_fun}{Filter function for the expression}

\item{...}{Further (metadata) attributes}
}
\description{
This block provides access to \link[dplyr]{filter} verb and
returns the filtered data.
}
