% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{new_stack}
\alias{new_stack}
\alias{is_stack}
\alias{generate_code.stack}
\alias{serve_stack}
\title{Stacks}
\usage{
new_stack(..., title = rand_names())

is_stack(x)

\method{generate_code}{stack}(x)

serve_stack(stack)
}
\arguments{
\item{...}{An ordered set of blocks (each argument is required to inherit
from \code{"block"})}

\item{title}{Stack title}

\item{x}{An object inheriting form \code{"stack"}}

\item{stack}{An object inheriting form \code{"stack"}}
}
\description{
A set of blocks can be assembled into a stack.
}
