% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{new_stack}
\alias{new_stack}
\alias{stack_server_generator}
\alias{serve_stack}
\title{Stacks}
\usage{
new_stack(..., name = rand_names())

stack_server_generator(stack)

serve_stack(stack)
}
\arguments{
\item{...}{An ordered set of blocks (each argument is required to inherit
from \code{"block"})}

\item{name}{Stack name}

\item{stack}{An object inheriting form \code{"stack"}}
}
\description{
A set of blocks can be assembled into a stack.
}
