% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{new_stack}
\alias{new_stack}
\alias{is_stack}
\alias{get_stack_name}
\alias{set_stack_name}
\alias{set_stack_title}
\alias{get_stack_title}
\alias{generate_code.stack}
\alias{serve_stack}
\title{Stacks}
\usage{
new_stack(..., title = "Stack", name = rand_names())

is_stack(x)

get_stack_name(x)

set_stack_name(x, name)

set_stack_title(x, title)

get_stack_title(x)

\method{generate_code}{stack}(x)

serve_stack(stack, id = "my_stack")
}
\arguments{
\item{...}{An ordered set of blocks (each argument is required to inherit
from \code{"block"})}

\item{title}{Stack title}

\item{name}{Stack name}

\item{x}{An object inheriting form \code{"stack"}}

\item{stack}{An object inheriting form \code{"stack"}}

\item{id}{Stack ID}
}
\description{
A set of blocks can be assembled into a stack.
}
