% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_numeric_field}
\alias{new_numeric_field}
\alias{numeric_field}
\title{Numeric field constructor}
\usage{
new_numeric_field(value = numeric(), min = numeric(), max = numeric(), ...)

numeric_field(...)
}
\arguments{
\item{value}{Field value}

\item{min, max}{Slider boundaries (inclusive)}

\item{...}{Further field components}
}
\description{
Numeric fields are translated into \link[shiny]{numericInput}
}
