% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{reset_shiny_outputs}
\alias{reset_shiny_outputs}
\title{Cleanup module outputs}
\usage{
reset_shiny_outputs(.output, session = getDefaultReactiveDomain())
}
\arguments{
\item{.output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\description{
Cleanup module outputs
}
\keyword{internal}
