% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_app}
\alias{run_app}
\title{Run app}
\usage{
run_app(..., extensions = new_dag_extension(), id = rand_names())
}
\arguments{
\item{..., extensions}{Forwarded to \code{\link[blockr.dock:dock]{blockr.dock::new_dock_board()}}}

\item{id}{Board namespace ID}
}
\value{
Acting a a wrapper to \code{\link[blockr.core:serve]{blockr.core::serve()}}, the result of a call
to \code{\link[shiny:shinyApp]{shiny::shinyApp()}} is returned.
}
\description{
Run a DAG-board app.
}
\examples{
if (interactive()) {
  run_app(
    blocks = c(
      a = new_dataset_block("iris")
    )
  )
}

}
