% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_app}
\alias{run_app}
\title{Run app}
\usage{
run_app(..., extensions = new_dag_extension(), id = rand_names())
}
\arguments{
\item{..., extensions}{Forwarded to \code{\link[blockr.dock:dock]{blockr.dock::new_dock_board()}}}

\item{id}{Board namespace ID}
}
\description{
Run a DAG-board app.
}
