% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_app}
\alias{run_app}
\title{Run app}
\usage{
run_app(..., extensions = blockr.dag::new_dag_extension())
}
\arguments{
\item{..., extensions}{Forwarded to \code{\link[blockr.dock:dock]{blockr.dock::new_dock_board()}}}
}
\description{
Run a DAG-board app.
}
