% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_select_field}
\alias{new_select_field}
\alias{select_field}
\title{Select field constructor}
\usage{
new_select_field(
  value = character(),
  choices = character(),
  multiple = FALSE,
  ...
)

select_field(...)
}
\arguments{
\item{value}{Default text input value.}

\item{choices}{List of values to select from. If elements of the list are
named, then that name --- rather than the value --- is displayed to the
user. It's also possible to group related inputs by providing a named list
whose elements are (either named or unnamed) lists, vectors, or factors. In
this case, the outermost names will be used as the group labels (leveraging
the \verb{<optgroup>} HTML tag) for the elements in the respective sublist. See
the example section for a small demo of this feature.}

\item{multiple}{Is selection of multiple items allowed?}

\item{...}{Other parameters passed to \link{new_field} and may be needed
by \link{ui_input} to pass more options to the related shiny input.}
}
\description{
Select fields are translated into \link[shiny]{selectInput}
}
