% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_select_field}
\alias{new_select_field}
\alias{select_field}
\title{Select field constructor}
\usage{
new_select_field(
  value = character(),
  choices = character(),
  multiple = FALSE,
  ...
)

select_field(...)
}
\arguments{
\item{value}{Field value}

\item{choices}{Set of permissible values}

\item{multiple}{Allow multiple selections}

\item{...}{Further field components}
}
\description{
Select fields are translated into \link[shiny]{selectInput}
}
