% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{stack_header}
\alias{stack_header}
\alias{stack_header.stack}
\title{Stack header generic}
\usage{
stack_header(x, ...)

\method{stack_header}{stack}(x, title, ns, ...)
}
\arguments{
\item{x}{Object for which to generate UI components}

\item{...}{Generic consistency}
}
\description{
Generic for creating stack header.
}
