% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_switch_field}
\alias{new_switch_field}
\alias{switch_field}
\title{Switch field constructor}
\usage{
new_switch_field(value = FALSE, ...)

switch_field(...)
}
\arguments{
\item{value}{Default text input value.}

\item{...}{Other parameters passed to \link{new_field} and may be needed
by \link{ui_input} to pass more options to the related shiny input.}
}
\description{
Switch fields are translated into \link[bslib]{input_switch}
}
