% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{to_json}
\alias{to_json}
\alias{to_json.field}
\alias{from_json}
\title{JSON serialization}
\usage{
to_json(x, ...)

\method{to_json}{field}(x, ...)

from_json(x, ...)
}
\arguments{
\item{x}{Object to (de)serialize}

\item{...}{Forwarded to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}}
}
\description{
Object serialization as JSON.
}
