% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_upload_field}
\alias{new_upload_field}
\alias{upload_field}
\title{Upload field constructor}
\usage{
new_upload_field(value = character(), ...)

upload_field(...)
}
\arguments{
\item{value}{Field value}

\item{...}{Further field components}
}
\description{
Upload fields are translated into \link[shiny]{fileInput}
}
