% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R, R/keyvalue_field.R, R/validation.R
\name{validate_field.string_field}
\alias{validate_field.string_field}
\alias{validate_field.select_field}
\alias{validate_field.switch_field}
\alias{validate_field.numeric_field}
\alias{validate_field.upload_field}
\alias{validate_field.filesbrowser_field}
\alias{validate_field.variable_field}
\alias{validate_field.range_field}
\alias{validate_field.list_field}
\alias{validate_field.keyvalue_field}
\alias{validate_field}
\alias{validate_field.field}
\alias{validate_field.default}
\alias{is_valid}
\alias{is_valid.field}
\alias{validation_failure}
\title{Validate field generic}
\usage{
\method{validate_field}{string_field}(x)

\method{validate_field}{select_field}(x)

\method{validate_field}{switch_field}(x)

\method{validate_field}{numeric_field}(x)

\method{validate_field}{upload_field}(x)

\method{validate_field}{filesbrowser_field}(x)

\method{validate_field}{variable_field}(x)

\method{validate_field}{range_field}(x)

\method{validate_field}{list_field}(x)

\method{validate_field}{keyvalue_field}(x)

validate_field(x)

\method{validate_field}{field}(x)

\method{validate_field}{default}(x)

is_valid(x)

\method{is_valid}{field}(x)

validation_failure(..., class = character())
}
\arguments{
\item{x}{A Field}

\item{...}{Message components (forwarded to \code{\link[=paste0]{paste0()}})}

\item{class}{Condition class (will be a subclass of \code{validation_failure})}
}
\description{
Validator functions for fields.
}
