% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-core.R
\name{value}
\alias{value}
\alias{value<-}
\title{Get field attribute value}
\usage{
value(x, name = "value")

value(x, name = "value") <- value
}
\arguments{
\item{x}{An object inheriting form \code{"field"}}

\item{name}{Field component name}

\item{value}{Field value}
}
\value{
Field value

The field.
}
\description{
Get the field value attribute. If it is a function,
return the result attribute instead.
}
