% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_variable_field}
\alias{new_variable_field}
\alias{variable_field}
\title{Variable field constructor}
\usage{
new_variable_field(
  value = character(),
  field = character(),
  components = list(),
  ...
)

variable_field(...)
}
\arguments{
\item{value}{Field value}

\item{field}{Field type}

\item{components}{Variable list of field components}

\item{...}{Further field components}
}
\description{
Variable field is intended to conditionally display
different field based on a condition.
}
\note{
Currently broken. Don't use.
}
