% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{new_variable_field}
\alias{new_variable_field}
\title{Variable field constructor}
\usage{
new_variable_field(field = character(), components = list(), ...)
}
\arguments{
\item{field}{Field type}

\item{components}{Variable list of field components}

\item{...}{Other parameters passed to \link{new_field} and may be needed
by \link{ui_input} to pass more options to the related shiny input.}
}
\description{
Variable field is intended to conditionally display
different field based on a condition.
}
\note{
Currently broken. Don't use.
}
